fix theme glitch with next-themes
