+++ package.json
@@ -73,6 +73,7 @@
     "next": "^13.0.3",
     "next-redux-wrapper": "^7.0.5",
     "next-seo": "^5.4.0",
+    "next-themes": "^0.2.1",
     "nextjs-progressbar": "^0.0.14",
     "react": "^18.1.0",
     "react-canvas-confetti": "^1.3.0",
